Move image.py out of analysis to avoid matplotlib import on core import
charuco.py (re-exported on the core path) imported to_gray from
lensboy.analysis.image. Importing that submodule runs
lensboy/analysis/__init__.py, which eagerly imports plots.py -> matplotlib,
so every `import lensboy` transitively pulled in matplotlib and scipy.

to_gray/to_color are generic uint8 image helpers with no dependency on
analysis, so move image.py to the top-level lensboy package and update the
two import sites. Core code no longer touches the optional analysis
subpackage at load time.

Cold `import lensboy` (median of 12 fresh interpreters, arm64):
before 382 ms (94 matplotlib + 175 scipy modules) -> after 73 ms (0/0).
